make signup view use the form
