+++ api/views.py
@@ -2,6 +2,7 @@ from django.urls import reverse_lazy
 from django.shortcuts import render, HttpResponse, get_object_or_404
 from django.views.generic import CreateView
 from django.contrib.auth.forms import UserCreationForm
+from api.forms import ClimbUserCreationForm
 
 import json
 from pages.models import *
@@ -17,8 +18,8 @@ def index(req):
     return render(req, "pages/example.html", ctx)
 
 class SignUpView(CreateView):
-    form_class = UserCreationForm
-    success_url = reverse_lazy("login")
+    form_class = ClimbUserCreationForm
+    success_url = reverse_lazy("accounts:login")
     template_name = "registration/signup.html"
 
 
@@ -50,3 +51,16 @@ def render_template(req, object, id, template):
     print("request: {template} with {}", ctx["data"])
     return render(req, template, ctx)
 
+
+
+def register(request):
+    if request.method == 'POST':
+        form = CustomUserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)
+            return redirect('index')
+    else:
+        form = CustomUserCreationForm()
+    
+    return render(request, 'registration/register.html', {'form': form})
